spectral: Move exposed members up a level
