+++ src/eradiate/spectral/__init__.pyi
@@ -1,3 +1,6 @@
-from . import ckd as ckd
-from . import index as index
-from . import mono as mono
+from .ckd import Bin as Bin
+from .ckd import BinSet as BinSet
+from .index import CKDSpectralIndex as CKDSpectralIndex
+from .index import MonoSpectralIndex as MonoSpectralIndex
+from .index import SpectralIndex as SpectralIndex
+from .mono import WavelengthSet as WavelengthSet
